requirements.txt
fastapi
uvicorn
google-cloud-discoveryengine
google-auth